requirements.txt
# Core Python dependencies
# Requires Python 3.12+

# Async and Performance
asyncio
uvloop>=0.19.0
aiohttp>=3.9.0
aiokafka>=0.10.0

# Data Processing and Validation
pydantic>=2.5.0
numpy>=1.26.0
scipy>=1.11.0

# GTFS-RT Protocol Buffers
protobuf>=4.25.0
gtfs-realtime-bindings>=1.0.0

# Database and Spatial
asyncpg>=0.29.0
geoalchemy2>=0.14.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9

# Logging and Monitoring
structlog>=23.3.0
prometheus-client>=0.19.0

# Configuration
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
